chore(testing): relocate MockAlphabet to test/utils per docs/tests/cipher.md
Modified files:
- backend/src/shared/testing/mock-alphabet.ts -> backend/test/utils/mock-alphabet.ts - moved to match the path docs/tests/cipher.md specifies as canonical; only *.spec.ts files need to live under Jest's rootDir, MockAlphabet itself does not
- backend/src/shared/testing/mock-alphabet.spec.ts, visual-alphabet-contract.spec.ts - updated import path
- backend/src/cipher/preprocess.spec.ts - updated import path
- backend/tsconfig.build.json - removed the now-unneeded src/shared/testing exclude entry (that directory only holds *.spec.ts files, already excluded globally; the existing top-level "test" exclude covers the new location)
