+++ backend/tsconfig.build.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["node_modules", "test", "dist", "**/*spec.ts", "src/shared/testing"]
+  "exclude": ["node_modules", "test", "dist", "**/*spec.ts"]
 }
